Moved NDV classes in prep for moving Stats mergers.
