+++ standalone-metastore/src/main/java/org/apache/hadoop/hive/common/ndv/hll/HLLConstants.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
