+++ standalone-metastore/src/test/java/org/apache/hadoop/hive/common/ndv/hll/TestHyperLogLogSparse.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -31,25 +31,22 @@ import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
 @RunWith(value = Parameterized.class)
-public class TestHLLNoBias {
+public class TestHyperLogLogSparse {
 
-  // 1.5% tolerance for long range bias (when no bias enabled) and 5% when (no
-  // bias is disabled) and
-  // 0.5% for short range bias
-  private float noBiaslongRangeTolerance = 2.0f;
-  private float biasedlongRangeTolerance = 5.0f;
-  private float shortRangeTolerance = 0.5f;
+  // 5% tolerance for long range bias and 1% for short range bias
+  private float longRangeTolerance = 5.0f;
+  private float shortRangeTolerance = 1.0f;
 
   private int size;
 
-  public TestHLLNoBias(int n) {
+  public TestHyperLogLogSparse(int n) {
     this.size = n;
   }
 
   @Parameters
   public static Collection<Object[]> data() {
-    Object[][] data = new Object[][] { { 30000 }, { 41000 }, { 50000 }, { 60000 }, { 75000 },
-        { 80000 }, { 81920 } };
+    Object[][] data = new Object[][] { { 2 }, { 10 }, { 100 }, { 1000 }, { 10000 }, { 100000 },
+        { 1000000 } };
     return Arrays.asList(data);
   }
 
@@ -61,7 +58,7 @@ public class TestHLLNoBias {
     for (int i = 0; i < size; i++) {
       hll.addLong(rand.nextLong());
     }
-    double threshold = size > 40000 ? noBiaslongRangeTolerance : shortRangeTolerance;
+    double threshold = size > 40000 ? longRangeTolerance : shortRangeTolerance;
     double delta = threshold * size / 100;
     assertEquals((double) size, (double) hll.count(), delta);
   }
@@ -71,44 +68,14 @@ public class TestHLLNoBias {
     Random rand = new Random(size);
     HyperLogLog hll = HyperLogLog.builder().build();
     int unique = size / 2;
-    Set<Long> hashset = new HashSet<Long>();
+    Set<Long> hashset = new HashSet<>();
     for (int i = 0; i < size; i++) {
       long val = rand.nextInt(unique);
       hashset.add(val);
       hll.addLong(val);
     }
-    double threshold = size > 40000 ? noBiaslongRangeTolerance : shortRangeTolerance;
+    double threshold = size > 40000 ? longRangeTolerance : shortRangeTolerance;
     double delta = threshold * hashset.size() / 100;
     assertEquals((double) hashset.size(), (double) hll.count(), delta);
   }
-
-  @Test
-  public void testHLLNoBiasDisabled() {
-    Random rand = new Random(size);
-    HyperLogLog hll = HyperLogLog.builder().enableNoBias(false).build();
-    int size = 100;
-    for (int i = 0; i < size; i++) {
-      hll.addLong(rand.nextLong());
-    }
-    double threshold = size > 40000 ? biasedlongRangeTolerance : shortRangeTolerance;
-    double delta = threshold * size / 100;
-    assertEquals((double) size, (double) hll.count(), delta);
-  }
-
-  @Test
-  public void testHLLNoBiasDisabledHalfDistinct() {
-    Random rand = new Random(size);
-    HyperLogLog hll = HyperLogLog.builder().enableNoBias(false).build();
-    int unique = size / 2;
-    Set<Long> hashset = new HashSet<Long>();
-    for (int i = 0; i < size; i++) {
-      long val = rand.nextInt(unique);
-      hashset.add(val);
-      hll.addLong(val);
-    }
-    double threshold = size > 40000 ? biasedlongRangeTolerance : shortRangeTolerance;
-    double delta = threshold * hashset.size() / 100;
-    assertEquals((double) hashset.size(), (double) hll.count(), delta);
-  }
-
 }
